Tweak secrets provider registration
